Moved enrichment function to separate files
