+++ metaspace/engine/sm/engine/annotation_lithops/pipeline.py
@@ -3,15 +3,14 @@ from __future__ import annotations
 import logging
 from typing import List, Tuple, Optional, Dict, Any
 
-import random
 import numpy as np
 import pandas as pd
 from lithops.storage.utils import CloudObject
-from pandas import DataFrame
 
 from sm.engine.annotation.diagnostics import FdrDiagnosticBundle
 from sm.engine.annotation.imzml_reader import LithopsImzMLReader
 from sm.engine.annotation.isocalc_wrapper import IsocalcWrapper
+from sm.engine.annotation_lithops.run_enrichment import run_enrichment
 from sm.engine.annotation_lithops.annotate import process_centr_segments
 from sm.engine.annotation_lithops.build_moldb import InputMolDb, DbFDRData
 from sm.engine.annotation_lithops.cache import PipelineCacher, use_pipeline_cache
@@ -33,9 +32,6 @@ from sm.engine.config import SMConfig
 from sm.engine.db import DB
 from sm.engine.ds_config import DSConfig
 
-from sm.engine import enrichment_db_molecule_mapping
-
-from sm.engine.errors import SMError
 
 logger = logging.getLogger('annotation-pipeline')
 
@@ -94,7 +90,7 @@ class Pipeline:  # pylint: disable=too-many-instance-attributes
 
     def run_pipeline(
         self, debug_validate=False, use_cache=True, perform_enrichment=False
-    ) -> Tuple[Dict[int, DataFrame], List[CObj[List[Tuple[int, bytes]]]], Any]:
+    ) -> Tuple[Dict[int, pd.DataFrame], List[CObj[List[Tuple[int, bytes]]]], Any]:
         # pylint: disable=unexpected-keyword-arg
         self.prepare_moldb(debug_validate=debug_validate)
 
@@ -194,52 +190,7 @@ class Pipeline:  # pylint: disable=too-many-instance-attributes
 
     @use_pipeline_cache
     def run_enrichment(self):
-        bootstrap_hash = {}
-        for moldb_id in self.results_dfs:
-            molecules = self.results_dfs[moldb_id]
-            data = []
-            for _, row in molecules.iterrows():
-                try:
-                    terms = enrichment_db_molecule_mapping.get_mappings_by_formula(
-                        row['formula'], str(moldb_id)
-                    )
-                    for term in terms:
-                        ion = str(row['formula']) + (
-                            str(row['modifier']) if row['modifier'] is not np.nan else ''
-                        )
-                        data.append([ion, term.molecule_enriched_name, term.id])
-                except SMError:
-                    pass
-            dataset_metabolites = pd.DataFrame(
-                data, columns=['formula_adduct', 'molecule_name', 'term_id']
-            )
-
-            # bootstrapping
-            hash_mol = {}
-            for _, row in dataset_metabolites.iterrows():
-                if row['formula_adduct'] not in hash_mol.keys():
-                    hash_mol[row['formula_adduct']] = []
-                hash_mol[row['formula_adduct']].append(
-                    {'mol_name': row['molecule_name'], 'mol_id': row['term_id']}
-                )
-
-            boot_n = 100  # times bootstrapping
-            bootstrap_sublist = []
-            for n in range(boot_n):
-                bootstrap_dict_aux = {}
-                for _, row in molecules.iterrows():
-                    ion = str(row['formula']) + str(row['modifier'])
-                    if ion in hash_mol.keys():
-                        bootstrap_dict_aux[ion] = random.choice(hash_mol[ion])
-                        bootstrap_sublist.append(
-                            [n, ion, row['fdr'], bootstrap_dict_aux[ion]['mol_id']]
-                        )
-
-            bootstrap_hash[moldb_id] = pd.DataFrame(
-                bootstrap_sublist,
-                columns=['scenario', 'formula_adduct', 'fdr', 'enrichment_db_molecule_mapping_id'],
-            )
-        self.enrichment_data = bootstrap_hash
+        self.enrichment_data = run_enrichment(self.results_dfs)
 
     @use_pipeline_cache
     def prepare_results(self):
