+++ metaspace/engine/sm/engine/annotation_lithops/run_enrichment.py
@@ -1,0 +1,65 @@
+from typing import Dict
+import logging
+import random
+import numpy as np
+import pandas as pd
+from sm.engine import enrichment_db_molecule_mapping
+from sm.engine.errors import SMError
+
+logger = logging.getLogger('annotation-pipeline')
+
+
+def run_enrichment(results_dfs: Dict[int, pd.DataFrame]):
+    bootstrap_hash = {}
+    for moldb_id in results_dfs:
+        print(f'Building enrichment data for moldb id {moldb_id}')
+        molecules = results_dfs[moldb_id]
+        data = []
+        for _, row in molecules.iterrows():
+            try:
+                terms = enrichment_db_molecule_mapping.get_mappings_by_formula(
+                    row['formula'], str(moldb_id)
+                )
+                for term in terms:
+                    ion = str(row['formula']) + (
+                        str(row['modifier']) if row['modifier'] is not np.nan else ''
+                    )
+                    data.append([ion, term.molecule_enriched_name, term.id])
+            except SMError:
+                pass
+        dataset_metabolites = pd.DataFrame(
+            data, columns=['formula_adduct', 'molecule_name', 'term_id']
+        )
+
+        logger.info('Molecular mappings loaded...')
+
+        # bootstrapping
+        hash_mol = {}
+        for _, row in dataset_metabolites.iterrows():
+            if row['formula_adduct'] not in hash_mol.keys():
+                hash_mol[row['formula_adduct']] = []
+            hash_mol[row['formula_adduct']].append(
+                {'mol_name': row['molecule_name'], 'mol_id': row['term_id']}
+            )
+
+        boot_n = 100  # times bootstrapping
+        bootstrap_sublist = []
+        logger.info('Bootstraping scenarios...')
+        for n in range(boot_n):
+            bootstrap_dict_aux = {}
+            for _, row in molecules.iterrows():
+                ion = str(row['formula']) + str(row['modifier'])
+                if ion in hash_mol.keys():
+                    bootstrap_dict_aux[ion] = random.choice(hash_mol[ion])
+                    bootstrap_sublist.append(
+                        [n, ion, row['fdr'], bootstrap_dict_aux[ion]['mol_id']]
+                    )
+
+        logger.info('Bootstraping generated...')
+
+        bootstrap_hash[moldb_id] = pd.DataFrame(
+            bootstrap_sublist,
+            columns=['scenario', 'formula_adduct', 'fdr', 'enrichment_db_molecule_mapping_id'],
+        )
+
+    return bootstrap_hash
